requirements.txt
# TODO: The requirements.txt is present, but it’s unclear if it’s up to date or if all dependencies are pinned.
arabic-reshaper==3.0.0
asttokens==3.0.0
audioread==3.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2022.12.7
cffi==2.0.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.2.3
debugpy==1.8.16
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.1
filelock==3.19.1
Flask==2.2.4
fsspec==2025.9.0
gunicorn==20.1.0
huggingface-hub==0.34.5
idna==3.4
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.2
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.43.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.7
mpmath==1.3.0
msgpack==1.1.1
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numba==0.60.0
numpy==2.0.2
opencv-python==4.12.0.88
packaging==25.0
parso==0.8.5
platformdirs==4.4.0
pooch==1.8.2
prompt_toolkit==3.0.52
psutil==7.0.0
pure_eval==0.2.3
pycparser==2.23
pydub==0.25.1
Pygments==2.19.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pywin32==311
PyYAML==6.0.2
pyzmq==27.1.0
regex==2025.9.1
requests==2.29.0
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
soundfile==0.13.1
soupsieve==2.4.1
soxr==1.0.0
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.1
typing_extensions==4.15.0
urllib3==1.26.15
wcwidth==0.2.13
Werkzeug==2.2.3
wikipedia==1.4.0
zipp==3.23.0
